v1.0.33
Hintergrundfunktionen:
+ Änderung: Text "Sprachauswahl" ist keine Frage mehr. Anpassung der Funktion erfolgt später: Die Sprache soll zwar änderbar sein, jedoch erfolgt die Einstellung aus Omni heraus.
+ Änderung: Es werden Standardmäßig nur Aufträge mit dem Status "In Bearbeitung" angezeigt und nicht mehr alle. Dies kann über die Filterfunktion geändert werden.
+ Neue Funktion: Neue Einstellung: "Temporäre Berichtdateien löschen". Löscht alle editierbaren *_.pdf und *.pdf Dateien aus den Berichtverzeichnissen und behält nur die *_final.pdf* Dateien nach Fertigstellung. Ein Abschalten belässt diese Berichtdateien im Pfad.
+ Fehlerbehebung: InfoBar: "geöffneten Bericht wieder schließen" und Fortschrittverhinderung wenn dieser noch geöffnet ist. Dies verhindert ein "Steckenbleiben" wegen einer Zugriffsverweigerung auf die geöffnete PDF Datei.
+ Fehlerbehebung: "Schritt-Icon's" im linken Menüband (Häuschen-Icon und 1,2,3 - Icons) sind nicht mehr zur Navigation vorgesehen. Dies verhindert ein Springen zwischen Schritten wenn der Kartenleser noch beschäftigt ist.
+ Fehlerbehebung: Solange ein Bericht geöffnet ist, kann nicht von Schritt 2 auf Schritt 3 gewechselt werden.
+ Fehlerbehebung: Ein Prüfvorgang kann nicht gestartet werden, wenn der Bericht bereits exitiert und nicht überschrieben werden kann oder werden soll.

+++ CardCheckAssistant/ViewModels/Step3PageViewModel.cs
@@ -281,9 +281,13 @@ public class Step3PageViewModel : ObservableObject
             }
         }
 
-        catch (Exception e)
+        catch (Exception ex)
         {
-            LogWriter.CreateLogEntry(e);
+            await App.MainRoot.MessageDialogAsync(
+                "Fehler",
+                string.Format("Bitte melde den folgenden Fehler an mich:\n{0}", ex.Message));
+
+            LogWriter.CreateLogEntry(ex);
         }
     }
 
@@ -300,16 +304,20 @@ public class Step3PageViewModel : ObservableObject
             await ExecuteRFIDGearCommand();
             NextStepCanExecute = true;
         }
-        catch (Exception e)
+        catch (Exception ex)
         {
-            LogWriter.CreateLogEntry(e);
+            await App.MainRoot.MessageDialogAsync(
+                "Fehler",
+                string.Format("Bitte melde den folgenden Fehler an mich:\n{0}", ex.Message));
+
+            LogWriter.CreateLogEntry(ex);
         }
     }
 
     /// <summary>
     /// 
     /// </summary>
-    private void OpenReportCommand_Executed()
+    private async void OpenReportCommand_Executed()
     {
         try
         {
@@ -333,16 +341,20 @@ public class Step3PageViewModel : ObservableObject
 
             p.Start();
         }
-        catch (Exception e)
+        catch (Exception ex)
         {
-            LogWriter.CreateLogEntry(e);
+            await App.MainRoot.MessageDialogAsync(
+                "Fehler",
+                string.Format("Bitte melde den folgenden Fehler an mich:\n{0}", ex.Message));
+
+            LogWriter.CreateLogEntry(ex);
         }
     }
 
     /// <summary>
     /// 
     /// </summary>
-    private void OpenReportPathCommand_Executed()
+    private async void OpenReportPathCommand_Executed()
     {
         try
         {
@@ -363,9 +375,13 @@ public class Step3PageViewModel : ObservableObject
 
             p.Start();
         }
-        catch (Exception e)
+        catch (Exception ex)
         {
-            LogWriter.CreateLogEntry(e);
+            await App.MainRoot.MessageDialogAsync(
+                "Fehler",
+                string.Format("Bitte melde den folgenden Fehler an mich:\n{0}", ex.Message));
+
+            LogWriter.CreateLogEntry(ex);
         }
     }
 
@@ -425,16 +441,34 @@ public class Step3PageViewModel : ObservableObject
 
             fileStream.Close();
 
-            File.Delete(preFinalPath);
-            File.Delete(semiFinalPath);
+            try
+            {
+                if(settings.DefaultSettings.RemoveTemporaryReportsIsEnabled == true)
+                {
+                    File.Delete(preFinalPath);
+                    File.Delete(semiFinalPath);
+                }
+            }
+            catch (Exception ex)
+            {
+                await App.MainRoot.MessageDialogAsync(
+                    "Fehler",
+                    string.Format("Bitte melde den folgenden Fehler an mich:\n{0}", ex.Message));
+
+                return;
+            }
 
             navigation.SetCurrentNavigationViewItem(nextpage);
             nextpage.IsEnabled = true;
             
         }
-        catch (Exception e)
+        catch (Exception ex)
         {
-            LogWriter.CreateLogEntry(e);
+            await App.MainRoot.MessageDialogAsync(
+                "Fehler",
+                string.Format("Bitte melde den folgenden Fehler an mich:\n{0}", ex.Message));
+
+            LogWriter.CreateLogEntry(ex);
         }
     }
